update MatcherMassOrdersTestSuite after review
