+++ it/src/test/scala/com/wavesplatform/it/transactions/MassTransferTransactionSuite.scala
@@ -1,146 +1,114 @@
 package com.wavesplatform.it.transactions
 
-import com.wavesplatform.it.api.AsyncHttpApi._
 import com.wavesplatform.it.util._
-import org.scalatest.CancelAfterFailure
-import scorex.account.AddressOrAlias
-import scorex.api.http.assets.SignedMassTransferRequest
-import scorex.crypto.encode.Base58
-import scorex.transaction.assets.MassTransferTransaction
-import scorex.transaction.assets.MassTransferTransaction.{ParsedTransfer, Transfer}
-import scorex.transaction.assets.TransferTransaction.MaxAttachmentSize
+import scorex.transaction.TransactionParser.TransactionType
+import com.wavesplatform.it.api.SyncHttpApi._
+import scorex.transaction.assets.MassTransferTransaction.Transfer
 
-import scala.concurrent.Await
-import scala.concurrent.Future.{sequence, traverse}
 import scala.concurrent.duration._
 
-class MassTransferTransactionSuite extends BaseTransactionSuite with CancelAfterFailure {
+class MassTransferTransactionSuite extends BaseTransactionSuite {
 
   private val Timeout = 2.minutes
-  private val AssetQuantity = 100.waves
-  private val TransferAmount = 5.waves
-  private val LeasingAmount = 5.waves
-  private val LeasingFee = 0.003.waves
-  private val TransferFee = 0.002.waves
-  private val IssueFee = 5.waves
+  private val assetQuantity = 100.waves
+  private val transferAmount = 5.waves
+  private val leasingAmount = 5.waves
+  private val leasingFee = 0.003.waves
+  private val transferFee = notMiner.settings.feesSettings.fees(TransactionType.TransferTransaction.id)(0).fee
+  private val issueFee = 1.waves
+  private val massTransferFeePerTransfer = notMiner.settings.feesSettings.fees(TransactionType.MassTransferTransaction.id)(0).fee
 
-  test("asset transfer changes asset balances and sender's.waves balance is decreased by fee") {
-    val f = for {
-      (balance1, eff1) <- notMiner.accountBalances(firstAddress)
-      (balance2, eff2) <- notMiner.accountBalances(secondAddress)
-      (balance3, eff3) <- notMiner.accountBalances(thirdAddress)
 
-      assetId <- sender.issue(firstAddress, "name", "description", AssetQuantity, 8, reissuable = false, IssueFee).map(_.id)
-      _ <- nodes.waitForHeightAraiseAndTxPresent(assetId)
+  test("asset mass transfer changes asset balances and sender's.waves balance is decreased by fee.") {
 
-      transfers = List(Transfer(secondAddress, TransferAmount), Transfer(thirdAddress, TransferAmount))
-      transferId <- sender.massTransfer(firstAddress, transfers, TransferFee, Some(assetId)).map(_.id)
-      _ <- nodes.waitForHeightAraiseAndTxPresent(transferId)
-      _ <- notMiner.assertBalances(firstAddress, balance1 - TransferFee - IssueFee, eff1 - TransferFee - IssueFee)
-      _ <- notMiner.assertBalances(secondAddress, balance2, eff2)
-      _ <- notMiner.assertBalances(thirdAddress, balance3, eff3)
-      _ <- notMiner.assertAssetBalance(firstAddress, assetId, AssetQuantity - TransferAmount - TransferAmount)
-      _ <- notMiner.assertAssetBalance(secondAddress, assetId, TransferAmount)
-      _ <- notMiner.assertAssetBalance(thirdAddress, assetId, TransferAmount)
-    } yield succeed
+    val (balance1, eff1) = notMiner.accountBalances(firstAddress)
+    val (balance2, eff2) = notMiner.accountBalances(secondAddress)
 
-    Await.result(f, Timeout)
+    val transfers = List(Transfer(secondAddress, transferAmount))
+    val assetId = sender.issue(firstAddress, "name", "description", assetQuantity, 8, reissuable = false, issueFee).id
+    nodes.waitForHeightAraiseAndTxPresent(assetId)
+
+
+    val massTransferTransactionFee = calcFee(transfers.size)
+    val transferId = sender.massTransfer(firstAddress, transfers, massTransferTransactionFee, Some(assetId)).id
+    nodes.waitForHeightAraiseAndTxPresent(transferId)
+
+
+    notMiner.assertBalances(firstAddress, balance1 - massTransferTransactionFee - issueFee, eff1 - massTransferTransactionFee - issueFee)
+    notMiner.assertAssetBalance(firstAddress, assetId, assetQuantity - transferAmount)
+    notMiner.assertBalances(secondAddress, balance2, eff2)
+    notMiner.assertAssetBalance(secondAddress, assetId, transferAmount)
   }
 
-  test("waves transfer changes waves balances") {
-    val f = for {
-      (balance1, eff1) <- notMiner.accountBalances(firstAddress)
-      (balance2, eff2) <- notMiner.accountBalances(secondAddress)
-      (balance3, eff3) <- notMiner.accountBalances(thirdAddress)
+  test("waves mass transfer changes waves balances") {
 
-      transfers = List(Transfer(secondAddress, TransferAmount), Transfer(thirdAddress, TransferAmount))
-      transferId <- sender.massTransfer(firstAddress, transfers, TransferFee).map(_.id)
-      _ <- nodes.waitForHeightAraiseAndTxPresent(transferId)
-      _ <- notMiner.assertBalances(firstAddress,
-        balance1 - TransferAmount - TransferAmount - TransferFee, eff1 - TransferAmount - TransferAmount - TransferFee)
-      _ <- notMiner.assertBalances(secondAddress, balance2 + TransferAmount, eff2 + TransferAmount)
-      _ <- notMiner.assertBalances(thirdAddress, balance3 + TransferAmount, eff3 + TransferAmount)
-    } yield succeed
+    val (balance1, eff1) = notMiner.accountBalances(firstAddress)
+    val (balance2, eff2) = notMiner.accountBalances(secondAddress)
+    val (balance3, eff3) = notMiner.accountBalances(thirdAddress)
+    val transfers = List(Transfer(secondAddress, transferAmount), Transfer(thirdAddress, 2 * transferAmount))
 
-    Await.result(f, Timeout)
+    val massTransferTransactionFee = calcFee(transfers.size)
+    val transferId = sender.massTransfer(firstAddress, transfers, massTransferTransactionFee).id
+    nodes.waitForHeightAraiseAndTxPresent(transferId)
+
+    notMiner.assertBalances(firstAddress, balance1 - massTransferTransactionFee - 3 * transferAmount, eff1 - massTransferTransactionFee - 3 * transferAmount)
+    notMiner.assertBalances(secondAddress, balance2 + transferAmount, eff2 + transferAmount)
+    notMiner.assertBalances(thirdAddress, balance3 + 2 * transferAmount, eff3 + 2 * transferAmount)
   }
 
-  test("invalid transfer should not be in UTX or blockchain") {
-    import MassTransferTransaction.MaxTransferCount
-    val address2 = AddressOrAlias.fromString(secondAddress).right.get
-    val valid = MassTransferTransaction.create(
-      None, sender.privateKey,
-      List(ParsedTransfer(address2, TransferAmount)),
-      System.currentTimeMillis,
-      TransferFee, Array.emptyByteArray).right.get
-    val fromFuture = valid.copy(timestamp = valid.timestamp + 1.day.toMillis)
-    val tooManyTransfers = valid.copy(transfers = List.fill(MaxTransferCount + 1)(ParsedTransfer(address2, 1)))
-    val negativeTransfer = valid.copy(transfers = List(ParsedTransfer(address2, -1)))
-    val negativeFee = valid.copy(fee = 0)
-    val longAttachment = valid.copy(attachment = ("ab" * MaxAttachmentSize).getBytes)
+  test("can not make mass transfer without having enough waves") {
+    val (balance1, eff1) = notMiner.accountBalances(firstAddress)
+    val (balance2, eff2) = notMiner.accountBalances(secondAddress)
+    val transfers = List(Transfer(secondAddress, balance1 / 2), Transfer(thirdAddress, balance1 / 2))
 
-    val invalidTransfers = Seq(fromFuture, tooManyTransfers, negativeTransfer, negativeFee, longAttachment)
-    for (tx <- invalidTransfers) {
-      val id = tx.id()
-      val req = createSignedMassTransferRequest(tx)
-      val f = for {
-        _ <- assertBadRequest(sender.signedMassTransfer(req))
-        _ <- sequence(nodes.map(_.ensureTxDoesntExist(id.base58)))
-      } yield succeed
+    assertBadRequest2(sender.massTransfer(firstAddress, transfers, calcFee(transfers.size)))
 
-      Await.result(f, Timeout)
-    }
+    nodes.waitForHeightAraise()
+    notMiner.assertBalances(firstAddress, balance1, eff1)
+    notMiner.assertBalances(secondAddress, balance2, eff2)
   }
 
-  test("can not make transfer without having enough waves") {
-    val f = for {
-      fb <- traverse(nodes)(_.height).map(_.min)
-      (balance1, eff1) <- notMiner.accountBalances(firstAddress)
-      (balance2, eff2) <- notMiner.accountBalances(secondAddress)
-      (balance3, eff3) <- notMiner.accountBalances(thirdAddress)
+  test("try to make mass transfer when fee less then mininal ") {
 
-      transfers = List(Transfer(secondAddress, balance1 / 2), Transfer(thirdAddress, balance1 / 2))
-      transferFailureAssertion <- assertBadRequest(sender.massTransfer(firstAddress, transfers, TransferFee))
+    val (balance1, eff1) = notMiner.accountBalances(firstAddress)
+    val (balance2, eff2) = notMiner.accountBalances(secondAddress)
+    val transfers = List(Transfer(secondAddress, transferAmount))
 
-      _ <- traverse(nodes)(_.waitForHeight(fb + 2))
-      _ <- notMiner.assertBalances(firstAddress, balance1, eff1)
-      _ <- notMiner.assertBalances(secondAddress, balance2, eff2)
-      _ <- notMiner.assertBalances(thirdAddress, balance3, eff3)
-    } yield transferFailureAssertion
-
-    Await.result(f, Timeout)
+    assertBadRequest2(sender.massTransfer(firstAddress, transfers, transferFee))
+    nodes.waitForHeightAraise()
+    notMiner.assertBalances(firstAddress, balance1, eff1)
+    notMiner.assertBalances(secondAddress, balance2, eff2)
   }
 
-  test("can not make transfer without having enough of effective balance") {
-    val f = for {
-      fb <- traverse(nodes)(_.height).map(_.min)
-      (balance1, eff1) <- notMiner.accountBalances(firstAddress)
-      (balance2, eff2) <- notMiner.accountBalances(secondAddress)
+  test("can not make mass transfer without having enough of effective balance") {
+    val (balance1, eff1) = notMiner.accountBalances(firstAddress)
+    val (balance2, eff2) = notMiner.accountBalances(secondAddress)
+    val transfers = List(Transfer(secondAddress, balance1 - leasingFee - transferFee))
 
-      leaseTxId <- sender.lease(firstAddress, secondAddress, LeasingAmount, LeasingFee).map(_.id)
-      _ <- nodes.waitForHeightAraiseAndTxPresent(leaseTxId)
+    val leaseTxId = sender.lease(firstAddress, secondAddress, leasingAmount, leasingFee).id
+    nodes.waitForHeightAraiseAndTxPresent(leaseTxId)
 
-      transfers = List(Transfer(secondAddress, balance1 - LeasingFee - TransferFee))
-      transferFailureAssertion <- assertBadRequest(sender.massTransfer(firstAddress, transfers, TransferFee))
-
-      _ <- traverse(nodes)(_.waitForHeight(fb + 2))
-      _ <- notMiner.assertBalances(firstAddress, balance1 - LeasingFee, eff1 - LeasingAmount - LeasingFee)
-      _ <- notMiner.assertBalances(secondAddress, balance2, eff2 + LeasingAmount)
-    } yield transferFailureAssertion
-
-    Await.result(f, Timeout)
+    assertBadRequest2(sender.massTransfer(firstAddress, transfers, transferFee))
+    nodes.waitForHeightAraise()
+    notMiner.assertBalances(firstAddress, balance1 - leasingFee, eff1 - leasingAmount - leasingFee)
+    notMiner.assertBalances(secondAddress, balance2, eff2 + leasingAmount)
   }
 
-  private def createSignedMassTransferRequest(tx: MassTransferTransaction): SignedMassTransferRequest = {
-    import tx._
-    SignedMassTransferRequest(
-      Base58.encode(tx.sender.publicKey),
-      assetId.map(_.base58),
-      transfers.map { case ParsedTransfer(address, amount) => Transfer(address.stringRepr, amount) },
-      fee,
-      timestamp,
-      attachment.headOption.map(_ => Base58.encode(attachment)),
-      signature.base58
-    )
+  private def calcFee(numberOfRecipients: Int): Long = {
+    transferFee + numberOfRecipients * massTransferFeePerTransfer
   }
+
+  //
+  //  private def createSignedMassTransferRequest(tx: MassTransferTransaction): SignedMassTransferRequest = {
+  //    import tx._
+  //    SignedMassTransferRequest(
+  //      Base58.encode(tx.sender.publicKey),
+  //      assetId.map(_.base58),
+  //      transfers.map { case ParsedTransfer(address, amount) => Transfer(address.stringRepr, amount) },
+  //      fee,
+  //      timestamp,
+  //      attachment.headOption.map(_ => Base58.encode(attachment)),
+  //      signature.base58
+  //    )
+  //  }
 }
